# hotel-booking-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "description": "REST API cho hệ thống đặt phòng khách sạn.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/admin/dat-phong": {
            "get": {
                "tags": [
                    "Quản trị - Đặt phòng"
                ],
                "summary": "Lấy danh sách đặt phòng quản trị",
                "operationId": "bfc91fc89073a9c2038fac53ca1d8ada",
                "parameters": [
                    {
                        "name": "tim_kiem",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "trang_thai",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "trang_thai_thanh_toan",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách đặt phòng thành công"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    },
                    "403": {
                        "description": "Không có quyền quản trị"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dat-phong/{id}": {
            "get": {
                "tags": [
                    "Quản trị - Đặt phòng"
                ],
                "summary": "Lấy chi tiết đặt phòng quản trị",
                "operationId": "f39c6a528de3a535a5a3ec003172b07e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy chi tiết đặt phòng thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy đặt phòng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dat-phong/{id}/xac-nhan": {
            "put": {
                "tags": [
                    "Quản trị - Đặt phòng"
                ],
                "summary": "Xác nhận đặt phòng",
                "operationId": "aad1abbf7d36e5bc8f8710bc9be18a4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xác nhận đặt phòng thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy đặt phòng"
                    },
                    "422": {
                        "description": "Không thể xác nhận"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dat-phong/{id}/huy": {
            "put": {
                "tags": [
                    "Quản trị - Đặt phòng"
                ],
                "summary": "Hủy đặt phòng",
                "operationId": "82ea8914c7865014436caa08d5cef046",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hủy đặt phòng thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy đặt phòng"
                    },
                    "422": {
                        "description": "Đặt phòng đã hủy"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dat-phong/{id}/xac-nhan-thanh-toan": {
            "put": {
                "tags": [
                    "Quản trị - Đặt phòng"
                ],
                "summary": "Xác nhận thanh toán tại khách sạn",
                "operationId": "71a284651c17421cc73fcd2941cb0a35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xác nhận thanh toán thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy đặt phòng"
                    },
                    "422": {
                        "description": "Không thể xác nhận thanh toán"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/admin/phong": {
            "get": {
                "tags": [
                    "Quản trị - Phòng"
                ],
                "summary": "Lấy danh sách phòng quản trị",
                "operationId": "e2370dbe9e7419cb0c3eec323e3b8c6a",
                "parameters": [
                    {
                        "name": "tim_kiem",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách phòng thành công"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    },
                    "403": {
                        "description": "Không có quyền quản trị"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quản trị - Phòng"
                ],
                "summary": "Thêm phòng kèm ảnh",
                "operationId": "e9a6fb1a9823f9be8537d186aedbc87a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "so_phong",
                                    "loai_phong_id",
                                    "gia_phong",
                                    "so_nguoi_toi_da"
                                ],
                                "properties": {
                                    "so_phong": {
                                        "type": "string"
                                    },
                                    "loai_phong_id": {
                                        "type": "integer"
                                    },
                                    "ten_phong": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "gia_phong": {
                                        "type": "number"
                                    },
                                    "so_nguoi_toi_da": {
                                        "type": "integer"
                                    },
                                    "mo_ta": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "hinh_anh": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thêm phòng thành công"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/phong": {
            "get": {
                "tags": [
                    "Phòng"
                ],
                "summary": "Lấy danh sách phòng",
                "operationId": "c65fb6f9b8dab69da92e29c9fad163fc",
                "parameters": [
                    {
                        "name": "trang_thai",
                        "in": "query",
                        "description": "Lọc chính xác theo trạng thái phòng",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "loai_phong_id",
                        "in": "query",
                        "description": "Lọc chính xác theo ID loại phòng",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "so_nguoi",
                        "in": "query",
                        "description": "Lọc phòng có số người tối đa lớn hơn hoặc bằng giá trị này",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy dữ liệu thành công"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Tham số truy vấn không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dữ liệu không hợp lệ"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}